# 2swagger_collection.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "CSC309PP1-EndpointsTEST",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
  "components": {
    "securitySchemes": {
      "noauthAuth": {
        "type": "http",
        "scheme": "noauth"
      }
    }
  },
...
  "paths": {
...
    "/api/account/logout": {
      "post": {
        "tags": ["account"],
        "summary": "Logout",
        "requestBody": {
          "content": {}
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {}
            }
          }
        }
      }
    },
...
    "/api/bookings/user/flights-info": {
      "get": {
        "tags": ["bookings"],
        "summary": "Get-flights-info",
        "security": [
          {
            "noauthAuth": []
          }
        ],
        "parameters": [
          {
            "name": "lastName",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "Fill out the flight booking info you get from the previous itinerary",
            "example": "Liang"
          },
          {
            "name": "bookingReference",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "example": "85F4C1"
          },
          {
            "name": "itineraryId",
            "in": "query",
            "schema": {
              "type": "integer"
            },
            "example": "49"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {}
            }
          }
        }
      }
    },
    "/api/bookings/user/flight-suggests": {
      "get": {
        "tags": ["bookings"],
        "summary": "Get-flight-suggests",
        "security": [
          {
            "noauthAuth": []
          }
        ],
        "parameters": [
          {
            "name": "hotelId",
            "in": "query",
            "schema": {
              "type": "integer"
            },
            "example": "3"
          },
          {
            "name": "origin",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "example": "Toronto"
          },
          {
            "name": "date",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "example": "2024-11-18"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {}
            }
          }
        }
      }
    },
...
    "/api/bookings/user/itinerary/{id}": {
      "get": {
        "tags": ["bookings"],
        "summary": "Get-Itinerary-ByID",
        "security": [
          {
            "noauthAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "schema": {
              "type": "integer"
            },
            "required": true,
            "example": "39"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {}
            }
          }
        }
      },
      "delete": {
        "tags": ["bookings"],
        "summary": "Cancel-itinerary-ByID",
        "security": [
          {
            "noauthAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "schema": {
              "type": "integer"
            },
            "required": true,
            "example": "48"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {}
            }
          }
        }
      }
    },
...
    "/api/bookings/user/flight-cancel": {
      "delete": {
        "tags": ["bookings"],
        "summary": "Cancel-flights",
        "security": [
          {
            "noauthAuth": []
          }
        ],
        "parameters": [
          {
            "name": "lastName",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "example": "Liang"
          },
          {
            "name": "bookingReference",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "example": "85F4C1"
          },
          {
            "name": "itineraryId",
            "in": "query",
            "schema": {
              "type": "integer"
            },
            "example": "49"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {}
            }
          }
        }
      }
    },
...
  }
}